fix: swap PercentageStage 100% opt
