+++ packages/kit/src/views/Swap/pages/components/SwapQuoteInput.tsx
@@ -1,15 +1,11 @@
-import { memo, useCallback, useEffect, useMemo } from 'react';
+import { memo, useCallback, useEffect } from 'react';
 
-import BigNumber from 'bignumber.js';
-import { useIntl } from 'react-intl';
-
-import { IconButton, Stack, Toast, YStack } from '@onekeyhq/components';
+import { IconButton, Stack, YStack } from '@onekeyhq/components';
 import {
   useSwapActions,
   useSwapFromTokenAmountAtom,
   useSwapLimitPriceFromAmountAtom,
   useSwapLimitPriceToAmountAtom,
-  useSwapNativeTokenReserveGasAtom,
   useSwapQuoteCurrentSelectAtom,
   useSwapSelectFromTokenAtom,
   useSwapSelectToTokenAtom,
@@ -20,10 +16,7 @@ import {
   useSwapTypeSwitchAtom,
 } from '@onekeyhq/kit/src/states/jotai/contexts/swap';
 import { validateAmountInput } from '@onekeyhq/kit/src/utils/validateAmountInput';
-import { ETranslations } from '@onekeyhq/shared/src/locale';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
-import type { INumberFormatProps } from '@onekeyhq/shared/src/utils/numberUtils';
-import { numberFormat } from '@onekeyhq/shared/src/utils/numberUtils';
 import {
   checkWrappedTokenPair,
   equalTokenNoCaseSensitive,
@@ -46,14 +39,15 @@ interface ISwapQuoteInputProps {
   selectLoading?: boolean;
   onSelectToken: (type: ESwapDirectionType) => void;
   onSelectPercentageStage?: (stage: number) => void;
+  onBalanceMaxPress?: () => void;
 }
 
 const SwapQuoteInput = ({
   onSelectToken,
   selectLoading,
+  onBalanceMaxPress,
   onSelectPercentageStage,
 }: ISwapQuoteInputProps) => {
-  const intl = useIntl();
   const [fromInputAmount, setFromInputAmount] = useSwapFromTokenAmountAtom();
   const [toInputAmount, setToInputAmount] = useSwapToTokenAmountAtom();
   const swapQuoteLoading = useSwapQuoteLoading();
@@ -68,7 +62,6 @@ const SwapQuoteInput = ({
   const [swapLimitPriceFromAmount] = useSwapLimitPriceFromAmountAtom();
   const [swapLimitPriceToAmount] = useSwapLimitPriceToAmountAtom();
   const [swapTypeSwitchValue] = useSwapTypeSwitchAtom();
-  const [swapNativeTokenReserveGas] = useSwapNativeTokenReserveGasAtom();
   useSwapQuote();
   useSwapFromAccountNetworkSync();
 
@@ -153,59 +146,6 @@ const SwapQuoteInput = ({
     toToken,
   ]);
 
-  const reserveGasFormatter: INumberFormatProps = useMemo(() => {
-    return {
-      formatter: 'balance',
-      formatterOptions: {
-        tokenSymbol: fromToken?.symbol,
-      },
-    };
-  }, [fromToken?.symbol]);
-
-  const checkNativeTokenGasToast = useCallback(() => {
-    let maxAmount = new BigNumber(fromTokenBalance ?? 0);
-    if (fromToken?.isNative) {
-      const reserveGas = swapNativeTokenReserveGas.find(
-        (item) => item.networkId === fromToken.networkId,
-      )?.reserveGas;
-      if (reserveGas) {
-        maxAmount = BigNumber.max(
-          0,
-          maxAmount.minus(new BigNumber(reserveGas)),
-        ).decimalPlaces(fromToken?.decimals ?? 6, BigNumber.ROUND_DOWN);
-      }
-      let reserveGasFormatted: string | undefined | number = reserveGas;
-      if (reserveGas) {
-        reserveGasFormatted = numberFormat(
-          reserveGas.toString(),
-          reserveGasFormatter,
-        );
-      }
-      const message = intl.formatMessage(
-        {
-          id: reserveGasFormatted
-            ? ETranslations.swap_native_token_max_tip_already
-            : ETranslations.swap_native_token_max_tip,
-        },
-        {
-          num_token: reserveGasFormatted,
-        },
-      );
-      Toast.message({
-        title: message,
-      });
-    }
-    return maxAmount;
-  }, [
-    fromTokenBalance,
-    fromToken?.isNative,
-    fromToken?.networkId,
-    fromToken?.decimals,
-    swapNativeTokenReserveGas,
-    intl,
-    reserveGasFormatter,
-  ]);
-
   return (
     <YStack gap="$2">
       <SwapInputContainer
@@ -223,13 +163,7 @@ const SwapQuoteInput = ({
         }}
         onSelectPercentageStage={onSelectPercentageStage}
         amountValue={fromInputAmount.value}
-        onBalanceMaxPress={() => {
-          const maxAmount = checkNativeTokenGasToast();
-          setFromInputAmount({
-            value: maxAmount?.toFixed() ?? '',
-            isInput: true,
-          });
-        }}
+        onBalanceMaxPress={onBalanceMaxPress}
         onSelectToken={onSelectToken}
         balance={fromTokenBalance}
       />
